python/requirements.txt
annotated-types==0.7.0
anyio==4.10.0
blis==1.3.0
catalogue==2.0.10
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
cloudpathlib==0.21.1
confection==0.1.5
curated-tokenizers==0.0.9
curated-transformers==0.1.1
cymem==2.0.11
Deprecated==1.2.18
fastapi==0.116.1
filelock==3.19.1
fsspec==2025.7.0
fugashi==1.5.1
ginza==5.2.0
h11==0.16.0
hf-xet==1.1.7
huggingface-hub==0.34.4
idna==3.10
ja-ginza-bert-large @ https://github.com/megagonlabs/ginza/releases/download/v5.2.0/ja_ginza_bert_large-5.2.0b1-py3-none-any.whl#sha256=cd788f852701ace399c442bd96a37d81935678f4269491d5b4ba33ab5f199688
ja_core_news_trf @ https://github.com/explosion/spacy-models/releases/download/ja_core_news_trf-3.8.0/ja_core_news_trf-3.8.0-py3-none-any.whl#sha256=4c08436bb3b0792cd4fa3150b282a658ec66e2caf7a126901fc93c02afffa790
jaconv==0.4.0
Jinja2==3.1.6
langcodes==3.5.0
language_data==1.3.0
lingua-language-detector==2.1.1
marisa-trie==1.3.0
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
murmurhash==1.0.13
networkx==3.5
numpy==2.3.2
packaging==25.0
plac==1.4.5
preshed==3.0.10
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pykakasi==2.3.0
PyYAML==6.0.2
regex==2025.7.34
requests==2.32.4
rich==14.1.0
safetensors==0.6.2
shellingham==1.5.4
smart_open==7.3.0.post1
sniffio==1.3.1
spacy==3.8.7
spacy-alignments==0.9.2
spacy-curated-transformers==0.3.1
spacy-legacy==3.0.12
spacy-loggers==1.0.5
spacy-transformers==1.2.5
srsly==2.5.1
starlette==0.47.2
SudachiDict-core==20250515
SudachiPy==0.6.10
sympy==1.14.0
thinc==8.3.6
thinc_apple_ops==1.0.0
tokenizers==0.13.3
torch==2.8.0
tqdm==4.67.1
transformers==4.30.2
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.1
unidic-lite==1.0.8
urllib3==2.5.0
uvicorn==0.35.0
wasabi==1.1.3
weasel==0.4.1
wrapt==1.17.3
